feat: add one-liner wallet sync API with ElectrumSync and SyncOptions
