+++ crates/electrum/Cargo.toml
@@ -16,10 +16,14 @@ workspace = true
 [dependencies]
 bdk_core = { path = "../core", version = "0.6.1" }
 electrum-client = { version = "0.24.0", features = [ "proxy" ], default-features = false }
-bdk_chain = { path = "../chain", version = "0.23.2" }
+
 
 [dev-dependencies]
-bdk_testenv = { path = "../testenv" }
+
+bdk_chain = { path = "../chain" }
+anyhow = "1.0"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
 
 criterion = { version = "0.7" }
