+++ examples/example_one_liner/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "example_one_liner"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+bdk_chain = { path = "../../crates/chain", features = ["serde"] }
+bdk_electrum = { path = "../../crates/electrum" }
+bdk_core = { path = "../../crates/core" }
+electrum-client = { version = "0.24.0", features = ["proxy"], default-features = false }
+serde = { version = "1.0", features = ["derive"] }
